bco-tool/requirements_latest.txt
argcomplete==1.12.3
asgiref==3.5.2
attrs==23.1.0
backports.zoneinfo==0.2.1
bullet==2.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
coloredlogs==15.0.1
cryptography==36.0.2
Django==4.0.6
docker==5.0.3
dxpy==0.326.1
gunicorn==21.2.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.12.0
jsonschema==4.17.3
lark==1.1.2
miniwdl==1.6.0
packaging==23.1
pkgutil-resolve-name==1.3.10
psutil==5.9.1
pycparser==2.21
pygtail==0.12.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
PyYAML==6.0
regex==2022.7.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
websocket-client==0.54.0
xdg==5.1.1
zipp==3.8.1
